Graph populating on page, needs reconfiguring for db data and to sort data into pos and neg
