+++ package.json
@@ -12,12 +12,16 @@
   },
   "dependencies": {
     "axios": "^1.4.0",
+    "chart.js": "^4.3.2",
+    "chartjs-adapter-moment": "^1.0.1",
     "concurrently": "^8.2.0",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "moment": "^2.29.4",
     "pg": "^8.11.1",
     "react": "^18.2.0",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
     "recharts": "^2.7.2",
     "rsuite": "^5.37.1"
