+++ src/components/ReviewGraph/GraphForReviews.jsx
@@ -1,17 +1,23 @@
-import AllReviewsChart from './AllReviewsChart.jsx'
-import CustomBarChart from './CustomBarChart.jsx'
+import {useState, useEffect} from 'react'
+// import AllReviewsChart from './AllReviewsChart.jsx'
+import AllReviews from './AllReviews/AllReviews.jsx'
+// import RecentReviews from './RecentReviews/RecentReviews.jsx'
+import ReviewFilters from './ReviewFilters/ReviewFilters.jsx'
+
 
 const GraphForReviews = () => {
-    const reviewSumm = 'ChangeMe'
-    const tooltipText = 'ChangeMe'
-    const reviewNumText = 'changeMe'
-    const totalReviewCount = ''
-    const posReviewCount= ''
-    const negReviewCount = ''
-    const steamPurchaseCount = ''
-    const OtherPurchaseCount = ''
-    const englishReviews = ''
-    const yourLanguageReviewCount = ''
+
+    const [reviewData, setReviewData] = useState([])
+    useEffect(() => {
+        const getReviews = async () => {
+            const result = await fetch('https://steam-clone-zf6a.onrender.com/reviews')
+            const data = await result.json()
+            setReviewData(data)
+        }
+        getReviews()
+        
+    }, [])
+    console.log(reviewData)
     
 //testData
     const data = [
@@ -23,186 +29,16 @@ const GraphForReviews = () => {
 
 
 
+  
+    
     return (
         <>
-            <h2 className="user_reviews_header no_bottom_margin">Customer reviews</h2>
             <div id='graphForReviews'>
-                <div id="user_reviews_summary_bar">
-                    <div id="summary_section">
-                        <div id="title">Overall Reviews:</div>
-                        <span id="game_review_summary" className={reviewSumm} data-tooltip-html={tooltipText}>{reviewSumm}</span>
-                        <span id='reviewNum'>{reviewNumText}</span>
-                        <a className='tooltip' data-tooltip-text='This summary uses only reviews writtem by customers that purchase the game directly from Steam.'>
-                            <img src="https://store.akamai.steamstatic.com/public/shared/images/ico/icon_questionmark.png"></img>
-                        </a>
-                    </div>
-                </div>
-                <div id='review_histogram_rollup_container' className="review_histogram">
-                    <div id="review_histogram_rollup">
-                        <AllReviewsChart data={data} />
-                        <div className='flot-text'>
-                        </div>
-                        <div className="flot-y-axis flot-y1-axis yAxis y1Axis"></div>
-                        <canvas className='flot-overlay' width='585' height='250'></canvas>
-                    </div>
-                </div>
-                <div id="review_recent_container">
-                    <div id='recent_summary_bar'>
-                        <div className="summary_section">
-                            <div className="title">Overall Reviews:</div>
-                            <span id="recent_review_summary" className={reviewSumm} data-tooltip-html={tooltipText}>{reviewSumm}</span>
-                            <span id='recentReviewNum'>{reviewNumText}</span>
-                            <a className='tooltip' data-tooltip-text='This summary uses only reviews writtem by customers that purchase the game directly from Steam.'>
-                                <img src="https://store.akamai.steamstatic.com/public/shared/images/ico/icon_questionmark.png"></img>
-                            </a>
-                        </div>
-                    </div>
-                    <div id='recent_review_histogram'>
-                        <canvas className="flot-overlay" width="315" height="250"></canvas>
-                    </div>
-                </div>
-                <div id='reviewFilterOptions'>
-                    <div className='reviewsFilterMenu'>
-
-                        <div className='title'>Review type</div>
-                        <div className='reviewMenuFlyout'>
-                            <div className='reviewMenuFlyoutContent'>
-                                <input type="radio" name="review_type" value="all" id="review_type_all" checked="checked" />
-                                <label for='review_type_all'>"All&nbsp;"
-                                    <span className="userReviewCount">{totalReviewCount}</span>
-                                </label>
-
-                                <input type="radio" name="review_type" value="positive" id="review_type_positive" checked="checked" />
-                                <label for='review_type_positive'>"Positive&nbsp;"
-                                    <span className="userReviewCount">{posReviewCount}</span>
-                                </label>
-                                <input type="radio" name="review_type" value="negative" id="review_type_negative" checked="checked" />
-                                <label for='review_type_negative'>"Negative&nbsp;"
-                                    <span className="userReviewCount">{negReviewCount}</span>
-                                </label>
-                            </div>
-                        </div>
-                    </div>
-
-                    <div className='reviewsFilterMenu'>
-                        <div className='title'>Purchase type</div>
-                        <div className='reviewMenuFlyout'>
-                            <div className='reviewMenuFlyoutContent'>
-                                <input type="radio" name="purchase_type" value="all" id="purchase_type_all" checked="checked" />
-                                <label for='purchase_type_all'>"All&nbsp;"
-                                    <span className="userReviewCount">{totalReviewCount}</span>
-                                </label>
-
-                                <input type="radio" name="purchase_type" value="steam" id="purchase_type_steam" checked="checked" />
-                                <label for='purchase_type_steam'>"Steam&nbsp;"
-                                    <span className="userReviewCount">{steamPurchaseCount}</span>
-                                </label>
-                                <input type="radio" name="purchase_type" value="other" id="purchase_type_other" checked="checked" />
-                                <label for='purchase_type_other'>"Other&nbsp;"
-                                    <span className="userReviewCount">{OtherPurchaseCount}</span>
-                                </label>
-                            </div>
-                        </div>
-                    </div>
-                    <div className='reviewsFilterMenu'>
-                        <div className="title">Language</div>
-                        <div className="reviewMenuFlyoutContent">
-
-                            <input type="radio" name="review_language" id="review_language_all" value="all" checked="checked" />
-                            <label for="review_language_all">All Languages&nbsp;<span className="user_reviews_count">{englishReviews}</span>
-                            </label>
-                            <br />
-                            <input type="radio" name="review_language" id="review_language_mine" value="english" checked="" />
-                            <label for="review_language_mine">Your Languages&nbsp;
-                                <span className="user_reviews_count">{yourLanguageReviewCount}</span>
-                                <a className="tooltip" data-tooltip-html="languageTooltipText">
-                                    <img src="https://store.akamai.steamstatic.com/public/shared/images/ico/icon_questionmark_dark.png" />
-                                </a>
-                            </label>
-                            <br />
-                            <div className="user_reviews_customize_language">
-                                <a href="#">Customize</a>
-                            </div>
-                        </div>
-                    </div>
-                    <div className='reviewsFilterMenu'>
-                        <div className="title">Date Range</div>
-                        <div className='reviewMenuFlyoutContent'>
-                            <div className="user_reviews_date_range_explanation">
-                                To view reviews within a date range, please click and drag a selection on a graph above or click on a specific bar.							<br /><br />
-                                <span className="reviewFilterBtn" >
-                                    <span>Show graph</span>
-                                </span>
-                            </div>
-                            <input type="radio" name="review_date_range" id="review_date_range_all" value="all" checked="checked" />
-                            <label for="review_date_range_all">Lifetime</label><br />
-                            <input type="radio" name="review_date_range" id="review_date_range_histogram" value="include" checked="checked" disabled="disabled" />
-                            <label for="review_date_range_histogram">Only Specific Range (Select on graph above)&nbsp;</label><br />
-                            <input type="radio" name="review_date_range" id="review_date_range_exclude_histogram" value="exclude" disabled="disabled" />
-                            <label for="review_date_range_exclude_histogram">Exclude Specific Range (Select on graph above)&nbsp;</label><br />
-
-                        </div>
-
-                    </div>
-                    <div className='reviewsFilterMenu'>
-                        <div className="title">Playtime</div>
-                        <div className="reviewMenuFlyoutContent">
-
-                            <div className="user_reviews_steam_labs_desc">
-                                <a href="#">
-                                    <img src="https://cdn.akamai.steamstatic.com/store/labs/main/images/steam_labs_logo.svg" />
-                                    <span>Brought to you by Steam Labs</span>
-                                </a>
-                            </div>
-
-                            <div className="user_reviews_playtime_filter_explanation">
-                                Filter reviews by the user's playtime when the review was written:						</div>
-
-                            <input type="radio" name="review_playtime_preset" id="review_playtime_preset_0" value="0" checked="" />
-                            <label for="review_playtime_preset_0">No Minimum</label>
-                            <br />
-                            <input type="radio" name="review_playtime_preset" id="review_playtime_preset_1" value="1" />
-                            <label for="review_playtime_preset_1">Over 1 hour</label>
-                            <br />
-                            <input type="radio" name="review_playtime_preset" id="review_playtime_preset_10" value="10" />
-                            <label for="review_playtime_preset_10">Over 10 hours</label>
-                            <br />
-                            <input type="radio" name="review_playtime_preset" id="review_playtime_preset_100" value="100" />
-                            <label for="review_playtime_preset_100">Over 100 hours</label>
-                            <br />
-
-                            <div id="app_reviews_playtime_range_text">
-                                <span id="app_reviews_playtime_range_text_min">No minimum</span>
-                                to
-                                <span id="app_reviews_playtime_range_text_max">No maximum</span>
-                            </div>
-                            <input type="hidden" id="app_reviews_playtime_range_min" value="0" />
-                            <input type="hidden" id="app_reviews_playtime_range_max" value="0" />
-
-                        </div>
-                    </div>
-                    <div className="reviewDisplayAs">
-                        <span className="title">Display As: </span>
-                        <select id="review_context" >
-                            <option value="summary">Summary</option>
-                            <option value="all">Most Helpful</option>
-                            <option value="recent">Recent</option>
-                            <option value="funny">Funny</option>
-                        </select>
-                    </div>
-                    <div>
-                        <span id="review_show_graph_button">
-                            <span>Show graph</span> 
-                            <div className="graph_toggle_icon down">&nbsp;</div>
-                        </span>
-                        <span id="review_hide_graph_button">
-                            <span>Hide graph</span> 
-                            <div className="graph_toggle_icon up">&nbsp;</div>
-                        </span>
-                    </div>
-                </div>
+            <h2 className="user_reviews_header no_bottom_margin">Customer reviews</h2>
+                <AllReviews data={data} />
+                {/* <RecentReviews data={data} /> */}
+                <ReviewFilters />
             </div>
-
         </>
     )
 }
